Move FlatMappingSpliterator to common
