+++ common/src/main/java/com/scottlogic/deg/common/util/FlatMappingSpliterator.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.generator;
+package com.scottlogic.deg.common.util;
 
 import java.util.Objects;
 import java.util.Spliterator;
