+++ generator/src/main/java/com/scottlogic/deg/generator/decisiontree/TreeDecisionNode.java
@@ -1,10 +1,9 @@
 package com.scottlogic.deg.generator.decisiontree;
 
-import com.scottlogic.deg.generator.FlatMappingSpliterator;
+import com.scottlogic.deg.common.util.FlatMappingSpliterator;
 
 import java.util.*;
 import java.util.Objects;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
