+++ generator/src/main/java/com/scottlogic/deg/generator/generation/combinationstrategies/ReductiveCombinationStrategy.java
@@ -1,6 +1,6 @@
 package com.scottlogic.deg.generator.generation.combinationstrategies;
 
-import com.scottlogic.deg.generator.FlatMappingSpliterator;
+import com.scottlogic.deg.common.util.FlatMappingSpliterator;
 import com.scottlogic.deg.generator.generation.databags.DataBag;
 import com.scottlogic.deg.generator.utils.RestartableIterator;
