+++ generator/src/main/java/com/scottlogic/deg/generator/generation/databags/DataBag.java
@@ -3,7 +3,7 @@ package com.scottlogic.deg.generator.generation.databags;
 import com.scottlogic.deg.common.output.CellSource;
 import com.scottlogic.deg.common.output.RowSource;
 import com.scottlogic.deg.common.profile.Field;
-import com.scottlogic.deg.generator.*;
+import com.scottlogic.deg.common.util.FlatMappingSpliterator;
 import com.scottlogic.deg.common.output.GeneratedObject;
 
 import java.util.*;
