+++ generator/src/main/java/com/scottlogic/deg/generator/walker/CartesianProductDecisionTreeWalker.java
@@ -2,7 +2,7 @@ package com.scottlogic.deg.generator.walker;
 
 import com.google.inject.Inject;
 import com.scottlogic.deg.common.profile.Field;
-import com.scottlogic.deg.generator.FlatMappingSpliterator;
+import com.scottlogic.deg.common.util.FlatMappingSpliterator;
 import com.scottlogic.deg.common.profile.ProfileFields;
 import com.scottlogic.deg.generator.decisiontree.ConstraintNode;
 import com.scottlogic.deg.generator.decisiontree.DecisionNode;
